[Kernel] Add missing data types and support for data type (de)serialization
This PR is part of delta-io/delta#1783. It adds additional data types supported by Delta Lake protocol that were missing from the interfaces PR delta-io/delta#1808.

It also adds serialization and deserailization of table schema represented as `StructType`.

UTs

Closes delta-io/delta#1842

+++ kernel/kernel-api/src/main/java/io/delta/kernel/types/ShortType.java
@@ -16,14 +16,12 @@
 package io.delta.kernel.types;
 
 /**
- * The data type representing {@code long} type values.
+ * The data type representing {@code short} type values.
  */
-public class LongType extends BasePrimitiveType
-{
-    public static final LongType INSTANCE = new LongType();
+public class ShortType extends BasePrimitiveType {
+    public static final ShortType INSTANCE = new ShortType();
 
-    private LongType()
-    {
-        super("long");
+    private ShortType() {
+        super("short");
     }
 }
